# docs/05-openapi.yaml
...
openapi: 3.1.1
info:
  title: LinguaDesk API
  description: Pre-release M034 contract containing implemented public capabilities, automatic-verification local-account registration, retained verification/password-recovery operations, browser cookie-session, and independent-client bearer operations. No compatibility or deprecation guarantee is implied.
  version: 0.1.0-m034
paths:
...
  /api/accounts/register:
    post:
      tags:
        - RegistrationEndpoints
      summary: Register a local account
      description: Creates an automatically verified local account when the normalized email is new and always returns the same sign-in-required acknowledgment for a valid duplicate. No authentication credential is issued.
      operationId: registerLocalAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
        required: true
      responses:
        "202":
          description: Accepted
          headers:
            Cache-Control:
              description: Always `no-store` for local-account mutation responses.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationAccepted"
        "400":
          description: Bad Request
          headers:
            Cache-Control:
              description: Always `no-store` for local-account mutation responses.
              required: true
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/RegistrationProblemDetails"
        "415":
          description: Unsupported Media Type
          headers:
            Cache-Control:
              description: Always `no-store` for local-account mutation responses.
              required: true
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/RegistrationProblemDetails"
        "503":
          description: Service Unavailable
          headers:
            Cache-Control:
              description: Always `no-store` for local-account mutation responses.
              required: true
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/RegistrationProblemDetails"
...
components:
  schemas:
...
    RegistrationAccepted:
      required:
        - status
      type: object
      properties:
        status:
          description: Always `signInRequired` for an accepted new or existing normalized email.
          $ref: "#/components/schemas/RegistrationStatus"
    RegistrationProblemDetails:
      required:
        - title
        - status
        - detail
        - category
        - correlationId
      type: object
      properties:
        type:
          type:
            - "null"
            - string
          description: RFC 9457 problem type reference.
        title:
          type: string
          description: Short, stable problem summary.
        status:
          type: integer
          description: HTTP status code for this occurrence.
          format: int32
        detail:
          type: string
          description: Safe explanation that never echoes submitted account values.
        category:
          type: string
          description: "LinguaDesk error category: `invalidRequest` or `availability`."
        correlationId:
          type: string
          description: Opaque request correlation identifier.
        errors:
          type:
            - "null"
            - object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field messages keyed only by `email` or `password`; present for field validation failures.
    RegistrationRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          maxLength: 254
          type: string
          description: Local account email address. Maximum 254 Unicode scalar values; surrounding whitespace is invalid.
          format: email
        password:
          maxLength: 128
          minLength: 15
          type: string
          description: Write-only password containing 15 to 128 well-formed Unicode scalar values.
          format: password
          writeOnly: true
    RegistrationStatus:
      enum:
        - signInRequired
      type: string
      description: Always `signInRequired` for an accepted new or existing normalized email.
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: Secure, HttpOnly, SameSite=Lax, host-scoped nonpersistent browser session cookie.
      name: __Host-LinguaDesk.Session
      in: cookie
    antiforgeryHeader:
      type: apiKey
      description: Opaque request token paired with the Secure, HttpOnly, SameSite=Strict __Host-LinguaDesk.Antiforgery cookie.
      name: X-LinguaDesk-Antiforgery
      in: header
    antiforgeryCookie:
      type: apiKey
      description: Secure, HttpOnly, SameSite=Strict host-scoped cookie paired with X-LinguaDesk-Antiforgery.
      name: __Host-LinguaDesk.Antiforgery
      in: cookie
    bearerAuth:
      type: http
      description: "Opaque ASP.NET Core Identity bearer access token presented as `Authorization: Bearer ...`. Tokens are never set as cookies or persisted by browser JavaScript."
      scheme: bearer
...